Fix memory fault
If the generator is NULL, the GenerateUniform or GenerateNormal api should create generator by initializing it to default state. Previously CreateGenerator is mandatory before calling Uniform or Normal. The application calls GenerateNormal directly without calling CreateGenerator, this is the cause for memory fault. Fixed the issue

Reuse streams_buffer if the seed is same as previous.